# v-poker-uni-app/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "V-Poker 2.0 API",
    "description": "V-Poker 扑克游戏全栈项目后端API文档。V3经济模型：单一货币（筹码），多级代理返佣，抽水3%。",
    "version": "3.0.0",
    "contact": {
      "name": "V-Poker Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "开发环境"
    },
    {
      "url": "https://api.yourdomain.com",
      "description": "生产环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "基于Cookie的会话认证，登录后自动设置"
      }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": ["个人资料"],
        "summary": "获取个人资料",
        "responses": { "200": { "description": "用户资料" } }
      },
      "patch": {
        "tags": ["个人资料"],
        "summary": "更新个人资料",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": { "type": "string" },
                  "avatar": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "更新成功" } }
      }
    },
    "/api/profile/devices": {
      "post": {
        "tags": ["个人资料"],
        "summary": "绑定设备",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deviceId": { "type": "string" },
                  "deviceName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "绑定成功" } }
      },
      "delete": {
        "tags": ["个人资料"],
        "summary": "解绑设备",
        "parameters": [
          { "name": "deviceId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "解绑成功" } }
      }
    },
    "/api/profile/password": {
      "post": {
        "tags": ["个人资料"],
        "summary": "修改密码",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["oldPassword", "newPassword"],
                "properties": {
                  "oldPassword": { "type": "string" },
                  "newPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "修改成功" }, "400": { "description": "原密码错误" } }
      }
    },
    "/api/profile/force-change-password": {
      "post": {
        "tags": ["个人资料"],
        "summary": "强制修改密码（首次登录）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newPassword"],
                "properties": {
                  "newPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "修改成功" } }
      }
    },
...
  },
...
}